Email AWS poc up and running
